Add a new rector to help with trait promotions
We have a frequently used method chain that comes from a trait. That
trait now includes a new method that provides a short cut to the
previous method chain. The test contains an example of the actual
scenario we have. The replacement method chain is only 2 methods right
now, but I could see this interface being easy to extend into an
arbitrary number of methods if required. I didn't want to complect the
rector with this possibility until we had a real-world scenario that
required it.